move ConversationCreationValues to new file
